requirements.txt
DateTime==4.4
dearpygui==1.6.2
path==15.1.0
python-dateutil==2.8.1
python-dotenv==0.19.2
vk==2.0.2
vk-api==11.9.8
vk-bot==0.6.1
vk-client==0.2.0
